add-registertable
