+++ .gitignore
@@ -1,3 +1,6 @@
+#hotkyou
+/lib/firebase/
+
 # Miscellaneous
 *.class
 *.log
