+++ pubspec.yaml
@@ -39,6 +39,7 @@ dependencies:
   firebase_core: ^2.13.1
   cloud_firestore: ^4.8.0
   provider: ^6.0.5
+  flutter_picker: ^2.1.0
 
 dev_dependencies:
   flutter_test:
